backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
PyJWT==1.7.1
django-cors-headers==3.5.0
joblib==0.15.1
scikit-learn==0.23.1
emoji==0.5.4
bs4==0.0.1
instaloader==4.4.4
nltk==3.5
pandas==1.0.5
regex==2020.6.8